Created Game Object Class & Base Logic
The GameObj.js contains the definition fo the game object class, containing a name, ID, image, and developer for each game. Will later add genres and possibly date of publishing. The GameObj.jsx contains the basic front-end display for a game object. These game object displays will be what the user sees when a game is on their list.