Implemented Game Objs Matching Search Results
When the user enters a query in search and hits enter or the search button, they wll recieve a list of games matching their result. Currently, only the names of the games show up correctly, and the image still doesn't appear. Working on fixing that right now. Relocated GameObj.js to be in a models folder within Frontend so that list.jsx can hydrate the data from the database into game objs.