Fixed Game Obj Display for Search
Game objects will now correctly display all elements, including image, rating, and genres. The issue was that we temporarily had it so that search.js would only retrieve the name and id of a game when matching it. I edited it to parse all elements, fixed naming issues in GameObj.jsx and List.jsx, and edited the appearance to be cleaner on List.jsx and styles.css. I also created a new method called parseDate in GameObj.jsx to parse the date from the DB and display it cleaner. We can make a Date class if desired to separate this concern.