refactor: rename message for duplicate game add
renamed from "errorMessage" to "duplicateMessage", since the "error" in the previous name is slightly misleading. Although the user is unable to add the same game to a list more than once and is considered an error, it is vague and doesn't convey the purpose and context as well as "duplicateMessage".